Fix wrong logo position on plymouth
Do not add "window" position to the logo/progress position. On a single
screen system, that position is always (0,0), but on multi-screen system
it's a position of the output. It seems that plymouth calls relevant
callbacks for each output ("window") separately, so there is no need to
add window position manually.

Fixes https://gitlab.freedesktop.org/plymouth/plymouth/-/issues/205