Added border to all shapes in order to blend bordering colors together correctly
The qubes-logo is now consisting of only three paths
(one for each color) which are perfectly aligned at their edges.
The Q-"comma" is now part of the lower right path, not a rectangle.

Mainly intended for tasks other than design, such as milling,
cutting, and so forth. Changes are unlikely to affect visual
representations.